Some quick little changes.
So apparently in order to lock you need to add this mutex library. I
added the library but it doesn't compile normally on the machines. It
looks like other people are having this problem on the forum.
If you add in -std=c++0x. It compiles fine. I changed the makefile for
now, but we'll probably want tto ask peter if that is ok to do or if
it's going to mess stuff up.

It looks like the user stuff is working, and I can't replicate that
error we were getting when opening an existing blank spreadsheet.